tests: split FIDO2 crypttab tests into fido2_test.go
Group the 4 FIDO2 crypttab boot tests (TestCrypttabFido2,
TestCrypttabFido2NoDevice, TestCtxAwareFido2CancelOnFallback,
TestCtxAwareFido2CancelMultiToken) in fido2_test.go so the file name
matches the test focus. TestSystemdFido2 stays in systemd_test.go
with its TestSystemdTPM2* siblings (grouped by enrollment style).
